feat: integrate Prometheus metrics, update environment management, and improve Docker setup
- Introduced Prometheus metrics via `/metrics` endpoint using `prom-client` for observability.
- Added Grafana and Prometheus configurations (commented) in `docker-compose.yaml` for monitoring stack integration.
- Refined `.gitignore` to include environment and Firebase template files.
- Replaced `.env.example` with `.env.template` to align with updated configurations.
- Enhanced Makefile with new tasks for Drizzle migration management and development builds.
- Updated Docker setup with multi-stage build optimizations and dedicated `Dockerfile.development`.
- Fixed environment replacement for Angular builds; updated `angular.json`.
- Improved error logging and database initialization in `server.js`.
- Replaced Prisma logic with Drizzle ORM in database client.
- Enhanced Firebase admin configuration to support multiple environments.

+++ src/config/database/client.js
@@ -1,24 +1,24 @@
 const postgres = require('postgres');
 const { drizzle } = require('drizzle-orm/postgres-js');
 const { readFileSync } = require('node:fs');
+const fs = require('fs');
 const { tb_tasks } = require('./schema');
 
 const client = postgres(process.env.DATABASE_URL);
-
 const database = drizzle(client);
 
-if (process.env.NODE_ENV === 'development') {
-  async function importTasks() {
-    const tasksData = JSON.parse(readFileSync('tasks.json', 'utf8'));
-    const formattedData = tasksData.map((task) => ({ ...task, dueDate: new Date(task.dueDate) }));
-    await database.delete(tb_tasks);
-    await database.insert(tb_tasks).values(formattedData);
-  }
+async function importTasks() {
+  const filePath = 'tasks.json';
+  if (!fs.existsSync(filePath)) return;
+  const tasksData = JSON.parse(readFileSync(filePath, 'utf8'));
 
+  await database.delete(tb_tasks);
+  const formattedData = tasksData.map((task) => ({ ...task, dueDate: new Date(task.dueDate) }));
+  await database.insert(tb_tasks).values(formattedData);
+}
+
+if (process.env.NODE_ENV === 'development') {
   importTasks()
-    .then(() => {
-      console.log('Import Complete!');
-    })
     .catch((err) => {
       console.error('Error importing tasks:', err);
     });
